project/requirements.txt
Pillow==9.5.0
numpy==1.24.2
django==4.2
celery==5.2.7
redis==4.5.4
loguru==0.7.0
